requirements.txt
# Flask Core
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# Database
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
alembic==1.12.0

# Redis and Caching
redis==5.0.1
Flask-Caching==2.1.0

# AI and ML Libraries
openai==1.3.5
requests==2.31.0
huggingface-hub==0.17.3
transformers==4.35.2
torch==2.2.0
sentence-transformers==2.2.2

# Authentication and Security
PyJWT==2.8.0
bcrypt==4.0.1
cryptography==41.0.7
passlib[bcrypt]==1.7.4

# Background Tasks
celery[redis]==5.3.4

# Web Server and WSGI
gunicorn==21.2.0
gevent==23.9.1

# API Framework (FastAPI for additional services)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2

# WebRTC and Media
aiortc==1.6.0
aiofiles==23.2.1

# Async Support
aioredis==2.0.1

# Data Processing
pandas==2.1.3
numpy==1.25.2
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
click==8.1.7
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.7
itsdangerous==2.1.2

# Date and Time
python-dateutil==2.8.2
pytz==2023.3

# File Processing
Pillow==10.0.1
openpyxl==3.1.2

# Networking and HTTP
urllib3==2.0.7
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4

# Development and Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-asyncio==0.21.1
black==23.10.1
flake8==6.1.0

# Monitoring and Logging
psutil==5.9.6

# Docker Integration
docker==6.1.3

# Retry and Error Handling
tenacity==8.2.3

# WebSocket Support
websockets==12.0

# System Integration
subprocess32==3.5.4; python_version < "3.3"